fix: repair broken test suite (conftest GCP mocks + 3 test bugs)
The test suite was completely broken — pytest crashed during collection
before any test could run. Three root causes:

1. conftest.py imported google.cloud.firestore to stub it, but the
   cryptography C extension is broken (pyo3_runtime.PanicException).
   Fixed by injecting mock modules directly into sys.modules instead
   of importing the real libraries, while preserving the google.*
   namespace packages that TensorFlow/protobuf depend on.

2. firestore.transactional decorator replaced real functions with
   MagicMock when firestore was mocked. Fixed by making the mock's
   transactional a pass-through lambda.

3. test_firestore_transactions used Mock(spec=firestore.Transaction)
   which fails when firestore.Transaction is itself a MagicMock.
   Fixed by using plain MagicMock().

Also fixed two minor bugs surfaced by the now-passing tests:
- compute_per_column_contributions returned numpy floats instead of
  Python floats (test_contributions assertion failure)
- update_job_status logged enum repr instead of value on Python 3.11+

Result: 122 passed, 14 skipped, 0 failures.

+++ evaluate/outliers.py
@@ -114,14 +114,14 @@ def compute_per_column_contributions(
         ]
     else:
         contributions = [
-            (column_names[i], (per_attr_losses[i] / total_loss) * 100.0)
+            (column_names[i], float((per_attr_losses[i] / total_loss) * 100.0))
             for i in range(len(per_attr_losses))
         ]
 
     # Normalize to exactly 100% (address potential floating-point rounding issues)
     total_pct = sum(pct for _, pct in contributions)
     contributions = [
-        (col, (pct / total_pct) * 100.0)
+        (col, float((pct / total_pct) * 100.0))
         for col, pct in contributions
     ]
